requirements.txt
# Core dependencies
requests>=2.28.0
pyyaml>=6.0
pandas>=1.5.0
openpyxl>=3.0.0

# Resume parsing
python-docx>=0.8.11
PyPDF2>=3.0.0

# Similarity matching (optional but recommended)
sentence-transformers>=2.2.0
torch>=2.0.0

# For web scraping fallbacks
beautifulsoup4>=4.11.0
lxml>=4.9.0
